+++ Backend/app/Http/Controllers/ProjectsController.php
@@ -4,6 +4,7 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 
 use App\Models\User;
+use App\Services\FileManager\CopyHelper;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use App\Services\Extractor;
@@ -44,6 +45,9 @@ class ProjectsController extends Controller
 	 */
 	public function store( Request $request , Guard $auth )
 	{
+		if ( !$request->has( 'type' ) )
+			return response()->json( array( 'error' => "Select template first" ) , 500 );
+
 		$oUser = $auth->user();
 
 		if ( !$request->input( 'name' ) )
@@ -51,13 +55,13 @@ class ProjectsController extends Controller
 
 		$aProject = Project::where( 'name' , '=' , $request->input( 'name' ) )->where( 'user_id' , '=' , $oUser->id )->get();
 
-		if ( count($aProject) > 0 )
+		if ( count( $aProject ) > 0 )
 			return response()->json( array( 'error' => 'Project name already exists' ) , 500 );
 
-		if( $request->has('folder_name') )
-			$aProject = Project::Where('location','LIKE','%'.$request->input('folder_name').'%')->where( 'user_id' , '=' , $oUser->id )->get();
+		if ( $request->has( 'folder_name' ) )
+			$aProject = Project::Where( 'location' , 'LIKE' , '%' . $request->input( 'folder_name' ) . '%' )->where( 'user_id' , '=' , $oUser->id )->get();
 
-		if ( count($aProject) > 0 )
+		if ( count( $aProject ) > 0 )
 			return response()->json( array( 'error' => 'Folder name already exists' ) , 500 );
 
 		$aInputData = $request->only( 'name' );
@@ -69,38 +73,48 @@ class ProjectsController extends Controller
 		if ( !$oProject )
 			return response()->json( array( 'error' => "Unable to create project" ) , 500 );
 
+		// Create project folder if not exists
 		$sProjectFolder = $request->input( 'folder_name' , $oProject->id );
 		$sDestinationPath = CLIENTS_BASE_PATH . $oUser->Organization->id . DIRECTORY_SEPARATOR . $oUser->id . DIRECTORY_SEPARATOR . $sProjectFolder . DIRECTORY_SEPARATOR;
-
-		// Create Destination Dir
 		if ( !file_exists( $sDestinationPath ) )
 			File::makeDirectory( $sDestinationPath , $permissions = intval( "0777" , 8 ) , true );
-
 		$oProject->location = $sProjectFolder . DIRECTORY_SEPARATOR;
 
-		// Extract preset's zip file in project_id folder
-		if ( $request->has( 'preset_id' ) ) {
-			$oPreset = Preset::find( $request->input( 'preset_id' ) );
-			if ( $oPreset ) {
-				$oProject->preset_id = $request->input( 'preset_id' );
-				$sSource = PRESETS_BASE_PATH . $oPreset->zip_location;
-				$oExtractor = new Extractor( $sSource , $sDestinationPath );
-				$oExtractor->extract( $sSource , $sDestinationPath );
-				if ( !$oExtractor ) {
-					$oProject->delete();
-					File::deleteDirectory( $sDestinationPath );
-					return response()->json( array( 'error' => "Unable to extract template" ) , 500 );
+		if ( $request->input( 'type' ) == "template" ) {
+			// Copy Preset folder in project_id folder
+			if ( $request->has( 'preset_id' ) ) {
+				$oPreset = Preset::find( $request->input( 'preset_id' ) );
+				if ( $oPreset ) {
+					$oProject->preset_id = $request->input( 'preset_id' );
+					$sSource = PRESETS_BASE_PATH . $oPreset->zip_location . DIRECTORY_SEPARATOR;
+					$oCopy = new CopyHelper();
+					$oCopy->copy( $sSource , $sDestinationPath );
+
+					/*$oExtractor = new Extractor( $sSource , $sDestinationPath );
+					$oExtractor->extract( $sSource , $sDestinationPath );
+					if ( !$oExtractor ) {
+						$oProject->delete();
+						File::deleteDirectory( $sDestinationPath );
+						return response()->json( array( 'error' => "Unable to extract template" ) , 500 );
+					}*/
 				}
 			}
+		} else if ( $request->input( 'type' ) == "zip" ) {
+			$this->extractZip( $request , $sDestinationPath );
+		} else if ( $request->input( 'type' ) == "blank" ){
+			file_put_contents($sDestinationPath."index.html"," ");
 		}
 
 		// Extract thumb
 		$oSnpShotGenerator = new SnapshotGenerator();
-		$sSrc = "Backend/clients/" . $oUser->Organization->id . "/" . $oUser->id . "/" . $sProjectFolder . "/index1.html";
-		$sDest = CLIENTS_BASE_PATH . $oUser->Organization->id . DIRECTORY_SEPARATOR . $oUser->id . DIRECTORY_SEPARATOR . $sProjectFolder . ".png";
-		$sThumbPath = $oSnpShotGenerator->getAndSavePreview( $sSrc , $sDest );
-		if ( $sThumbPath ) {
-			$oProject->thumb = $sProjectFolder . ".png";
+		$aFiles = $oSnpShotGenerator->getHtmlFile( $sDestinationPath );
+		if ( !empty( $aFiles ) AND count( $aFiles ) > 1 ) {
+			$sSrc = "Backend/clients/" . $oUser->Organization->id . "/" . $oUser->id . "/" . $sProjectFolder . "/" . $aFiles[ 1 ];
+			$sDest = CLIENTS_BASE_PATH . $oUser->Organization->id . DIRECTORY_SEPARATOR . $oUser->id . DIRECTORY_SEPARATOR . $sProjectFolder . ".png";
+			$sThumbPath = $oSnpShotGenerator->getAndSavePreview( $sSrc , $sDest );
+			if ( $sThumbPath ) {
+				$oProject->thumb = $sProjectFolder . ".png";
+			}
 		}
 
 		$oProject->save();
@@ -156,4 +170,34 @@ class ProjectsController extends Controller
 		//
 	}
 
+	private function extractZip(Request $request ,$sDestinationPath )
+	{
+		if( $request->hasFile('zipFile') ) {
+			$oFile = $request->file( 'zipFile' );
+			$sZipFile = $request->file( 'zipFile' )->getClientOriginalName();
+			$sFileName = pathinfo( $oFile->getClientOriginalName() , PATHINFO_FILENAME );
+
+			$counter = 1;
+			while ( File::exists( $sDestinationPath . $sFileName ) ) {
+				$sFileName = pathinfo( $oFile->getClientOriginalName() , PATHINFO_FILENAME ) . "_" . $counter ++;
+				$sZipFile = $sFileName . "." . pathinfo( $oFile->getClientOriginalName() , PATHINFO_EXTENSION );
+			}
+
+			$request->file( 'zipFile' )->move(  CLIENTS_BASE_PATH , $sZipFile );
+
+			// Extract zip destination folder
+			$sSourceZip = CLIENTS_BASE_PATH . $sZipFile;
+			$sDestinationFolder = $sDestinationPath;
+			if ( !file_exists( $sDestinationFolder ) )
+				File::makeDirectory( $sDestinationFolder , 0777 , true );
+			$oExtractor = new Extractor( $sSourceZip , $sDestinationFolder );
+			$oExtractor->extract( $sSourceZip , $sDestinationFolder );
+
+			unlink( $sSourceZip );
+
+			if( $oExtractor )
+				return true;
+		}
+		return false;
+	}
 }
